Use u8 for ipproto
In a few places in app layer and unittests u16 was used.

+++ src/app-layer-dns-common.c
@@ -69,7 +69,7 @@ int DNSStateGetEventInfo(const char *event_name,
     return 0;
 }
 
-void DNSAppLayerRegisterGetEventInfo(uint16_t ipproto, uint16_t alproto)
+void DNSAppLayerRegisterGetEventInfo(uint8_t ipproto, uint16_t alproto)
 {
     AppLayerParserRegisterGetEventInfo(ipproto, alproto, DNSStateGetEventInfo);
